refactor: 로또통계 객체 생성 시 당첨정보를 집계하도록 수정
